Fixed history issue, adding navbar
