+++ app/javascript/components/navbar/NavBar.jsx
@@ -1,0 +1,92 @@
+import React, { Component } from "react";
+import { Link } from "react-router-dom";
+import {
+  Button,
+  Container,
+  Menu,
+  Responsive,
+  Segment,
+  Visibility
+} from "semantic-ui-react";
+
+const getWidth = () => {
+  const isSSR = typeof window === "undefined";
+
+  return isSSR ? Responsive.onlyTablet.minWidth : window.innerWidth;
+};
+
+class NavBar extends Component {
+  state = {};
+
+  hideFixedMenu = () => this.setState({ fixed: false });
+  showFixedMenu = () => this.setState({ fixed: true });
+
+  render() {
+    const { children } = this.props;
+    const { fixed } = this.state;
+
+    return (
+      <div>
+        <Responsive
+          getWidth={getWidth}
+          minWidth={Responsive.onlyTablet.minWidth}
+        >
+          <Visibility
+            once={false}
+            onBottomPassed={this.showFixedMenu}
+            onBottomPassedReverse={this.hideFixedMenu}
+          >
+            <Segment
+              inverted
+              textAlign="center"
+              style={{ minHeight: 50, padding: "1em 0em" }}
+              vertical
+            >
+              <Menu
+                fixed={fixed ? "top" : null}
+                inverted={!fixed}
+                pointing={!fixed}
+                secondary={!fixed}
+                size="large"
+              >
+                <Container>
+                  <Menu.Item as={Link} to="/">
+                    Home
+                  </Menu.Item>
+                  <Menu.Item as={Link} to="/UserProfiles">User Profiles</Menu.Item>
+
+                  <Menu.Item position="right">
+                    <Button
+                      color="green"
+                      inverted={!fixed}
+                      size="large"
+                      as={Link}
+                      to="/login"
+                      icon="sign in alternate"
+                      content="Login"
+                    ></Button>
+                    <Button
+                      color="blue"
+                      size="large"
+                      as={Link}
+                      to="/signup"
+                      icon="registered"
+                      content="Sign up"
+                      inverted={!fixed}
+                      primary={fixed}
+                      style={{ marginLeft: "0.5em" }}
+                    ></Button>
+                  </Menu.Item>
+                </Container>
+              </Menu>
+            </Segment>
+          </Visibility>
+
+          {children}
+        </Responsive>
+      </div>
+    );
+  }
+}
+
+export default NavBar;
