+++ app/javascript/components/registrations/Signup.jsx
@@ -80,7 +80,7 @@ class Signup extends Component {
       <div>
         <Grid
           textAlign="center"
-          style={{ height: "100vh" }}
+          style={{ marginTop: "2em"}}
           verticalAlign="middle"
         >
           <Grid.Column style={{ maxWidth: 450 }}>
