+++ app/javascript/components/userprofiles/UserPage.jsx
@@ -1,0 +1,13 @@
+import React, { Component } from "react";
+
+class UserPage extends Component {
+
+ render() {
+     return (
+     <div>
+         <p>This is the list of users page</p>
+     </div>);
+ }
+}
+
+export default UserPage;
