+++ app/javascript/components/userprofiles/UserProfile.jsx
@@ -1,0 +1,13 @@
+import React, { Component } from "react";
+
+class UserProfile extends Component {
+
+ render() {
+     return (
+     <div>
+
+     </div>);
+ }
+}
+
+export default UserProfile;
