+++ app/javascript/packs/index.jsx
@@ -6,11 +6,14 @@ import ReactDOM from "react-dom";
 import { BrowserRouter as Router, Route } from "react-router-dom";
 import App from "../components/App";
 import "../../assets/dist/semantic.min.css";
+import { createBrowserHistory } from "history";
 
 document.addEventListener("DOMContentLoaded", () => {
+
+  const history = createBrowserHistory();
   ReactDOM.render(
-    <Router>
-      <Route path="/" component={App} />
+    <Router history = {history}>
+      <Route path="/" component={App}  />
     </Router>,
     document.body.appendChild(document.createElement("div"))
   );
